+++ src/cmd/shared.go
@@ -2,10 +2,7 @@ package cmd
 
 import (
 	"errors"
-	"fmt"
-	"os"
 
-	"github.com/git-town/git-town/src/drivers"
 	"github.com/git-town/git-town/src/git"
 	"github.com/git-town/git-town/src/prompt"
 	"github.com/git-town/git-town/src/steps"
@@ -42,51 +39,6 @@ func ValidateIsRepository(repo *git.ProdRepo) error {
 		return nil
 	}
 	return errors.New("this is not a Git repository")
-}
-
-func ensureIsNotInUnfinishedState(repo *git.ProdRepo, driver drivers.CodeHostingDriver) error {
-	runState, err := steps.LoadPreviousRunState(repo)
-	if err != nil {
-		return fmt.Errorf("cannot load previous run state: %w", err)
-	}
-	if runState == nil || !runState.IsUnfinished() {
-		return nil
-	}
-	response := prompt.AskHowToHandleUnfinishedRunState(
-		runState.Command,
-		runState.UnfinishedDetails.EndBranch,
-		runState.UnfinishedDetails.EndTime,
-		runState.UnfinishedDetails.CanSkip,
-	)
-	switch response {
-	case prompt.ResponseTypeDiscard:
-		return steps.DeletePreviousRunState(repo)
-	case prompt.ResponseTypeContinue:
-		hasConflicts, err := repo.Silent.HasConflicts()
-		if err != nil {
-			return err
-		}
-		if hasConflicts {
-			return fmt.Errorf("you must resolve the conflicts before continuing")
-		}
-		err = steps.Run(runState, repo, driver)
-		if err != nil {
-			return err
-		}
-	case prompt.ResponseTypeAbort:
-		abortRunState := runState.CreateAbortRunState()
-		err = steps.Run(&abortRunState, repo, driver)
-	case prompt.ResponseTypeSkip:
-		skipRunState := runState.CreateSkipRunState()
-		err = steps.Run(&skipRunState, repo, driver)
-	default:
-		return fmt.Errorf("unknown response: %s", response)
-	}
-	if err != nil {
-		return err
-	}
-	os.Exit(0)
-	return nil
 }
 
 func getAppendStepList(config appendConfig, repo *git.ProdRepo) (result steps.StepList, err error) {
